Check the permissions for destroy image in FileManager module

+++ Modules/FileManager/App/Services/ImageService.php
@@ -44,7 +44,7 @@ class ImageService
 
     public function destroy(Image $image): bool|null
     {
-//        Gate::authorize('destroy', $image);
+        Gate::authorize('destroy', $image);
         return $image->delete();
     }
